Исправить сборку публикаций
